requirements.txt
django==5.2.1
djangorestframework==3.16.0
okta-jwt-verifier==0.2.5
cryptography==45.0.2